PAXSB-17 automatically set version ranges of Fix-Package Header and calculate a default fix package version.
git-svn-id: https://scm.ops4j.org/repos/ops4j/projects/pax/swissbox@15640 9b982a3c-3ae5-0310-a4bc-d9a3335569bd

+++ pax-swissbox-tinybundles/src/main/java/org/ops4j/pax/swissbox/tinybundles/dp/store/TemporaryBinaryStore.java
@@ -46,8 +46,13 @@ public class TemporaryBinaryStore implements BinaryStore<InputStream>
 
     public TemporaryBinaryStore()
     {
+        this( false );
+    }
+
+    public TemporaryBinaryStore( boolean flushStoreage )
+    {
         m_dir = new File( System.getProperty( "java.io.tmpdir" ) + "/tb" );
-        if( m_dir.exists() )
+        if( m_dir.exists() && flushStoreage )
         {
             FileUtils.delete( m_dir );
         }
@@ -92,7 +97,7 @@ public class TemporaryBinaryStore implements BinaryStore<InputStream>
 
     private File getLocation( String id )
     {
-        return new File( m_dir, "tinyundles_" + id + ".bin" );
+        return new File( m_dir, "tinybundles_" + id + ".bin" );
     }
 
     public InputStream load( BinaryHandle handle )
